fix: codegen failing to find imported types
