+++ src/common/types.ts
@@ -1,124 +1,11 @@
-export interface OnChangeTextEvent {
-  value: string;
-}
-
-export interface OnChangeHtmlEvent {
-  value: string;
-}
-
-export interface OnChangeStateEvent {
-  bold: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  italic: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  underline: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  strikeThrough: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  inlineCode: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  h1: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  h2: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  h3: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  h4: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  h5: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  h6: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  codeBlock: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  blockQuote: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  orderedList: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  unorderedList: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  link: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  image: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-  mention: {
-    isActive: boolean;
-    isConflicting: boolean;
-    isBlocking: boolean;
-  };
-}
-
-export interface OnChangeStateDeprecatedEvent {
-  isBold: boolean;
-  isItalic: boolean;
-  isUnderline: boolean;
-  isStrikeThrough: boolean;
-  isInlineCode: boolean;
-  isH1: boolean;
-  isH2: boolean;
-  isH3: boolean;
-  isH4: boolean;
-  isH5: boolean;
-  isH6: boolean;
-  isCodeBlock: boolean;
-  isBlockQuote: boolean;
-  isOrderedList: boolean;
-  isUnorderedList: boolean;
-  isLink: boolean;
-  isImage: boolean;
-  isMention: boolean;
-}
+// Re-export event types from the NativeComponent spec file (source of truth for Codegen)
+export type {
+  OnChangeTextEvent,
+  OnChangeHtmlEvent,
+  OnChangeStateEvent,
+  OnChangeStateDeprecatedEvent,
+  OnKeyPressEvent,
+} from '../native/EnrichedTextInputNativeComponent';
 
 export interface OnMentionDetected {
   text: string;
@@ -142,10 +29,6 @@ export interface OnChangeSelectionEvent {
 export interface OnChangeMentionEvent {
   indicator: string;
   text: string;
-}
-
-export interface OnKeyPressEvent {
-  key: string;
 }
 
 export interface EnrichedTextInputInstanceBase {
